killing old DEV_PDF branch - will make a new one from trunk

+++ src/wordcram/Placers.java
@@ -18,8 +18,7 @@ package wordcram;
 
 import java.util.Random;
 
-import processing.core.PApplet;
-import processing.core.PVector;
+import processing.core.*;
 
 public class Placers {
 
@@ -76,7 +75,7 @@ public class Placers {
 			public PVector place(Word word, int wordIndex, int wordsCount,
 					int wordImageWidth, int wordImageHeight, int fieldWidth,
 					int fieldHeight) {
-				float x = (1 - word.weight) * fieldWidth * r.nextFloat(); // big=left, small=right
+				float x = (float) (1 - word.weight) * fieldWidth * r.nextFloat(); // big=left, small=right
 				float y = ((float) fieldHeight) * 0.5f;
 				return new PVector(x, y);
 			}
